Changed velocity of entity to float

+++ src/entity.h
@@ -6,8 +6,8 @@
 
 struct Velocity
 {
-	int x = 0;
-	int y = 0;
+	float x = 0.0f;
+	float y = 0.0f;
 };
 
 struct Entity
